ingestion/requirements.txt
requests
snowflake-connector-python
python-dotenv